# specs/001-quant-ops-dashboard/contracts/rest-api.openapi.yaml
openapi: 3.1.0
info:
  title: Kalshi Quant Dashboard API
  version: 0.3.0
  description: |
    Authenticated REST API for historical queries, live or historical convergence,
    admin-managed access policy, feature-flag mutation, alert-rule management,
    role-scoped diagnostics, and export enforcement.
servers:
  - url: https://internal.example.com
    description: Production internal deployment
  - url: http://localhost:3001
    description: Local development
security:
  - cookieAuth: []
paths:
...
  /api/overview:
    get:
      summary: Get overview dashboard snapshot
      operationId: getOverview
      parameters:
        - $ref: '#/components/parameters/Timezone'
      responses:
        '200':
          description: Overview snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/admin/alert-rules:
    get:
      summary: List persisted alert rule configuration
      operationId: listAlertRules
      responses:
        '200':
          description: Alert rule configs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRuleListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/alert-rules/{alertRuleId}:
    patch:
      summary: Update one alert rule threshold or enablement
      operationId: updateAlertRule
      parameters:
        - $ref: '#/components/parameters/AlertRuleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleUpdateRequest'
      responses:
        '200':
          description: Updated alert rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRuleMutationResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: kqd_session
  parameters:
...
    AlertRuleId:
      name: alertRuleId
      in: path
      required: true
      schema:
        type: string
...
    Timezone:
      name: timezone
      in: query
      schema:
        type: string
        enum: [utc, local]
        default: utc
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not authorized for the requested surface
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Version conflict or conflicting policy state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConflictResponse'
    ValidationFailed:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    OverviewResponse:
      type: object
      required:
        - generatedAt
        - healthSummary
        - aggregatePnl
        - liveDecisionFeed
        - liveTradeFeed
        - queueSummary
        - recentAlerts
      properties:
        generatedAt:
          type: string
          format: date-time
        healthSummary:
          $ref: '#/components/schemas/SystemHealthSummary'
        aggregatePnl:
          $ref: '#/components/schemas/PnlSummaryCard'
        liveDecisionFeed:
          type: array
          items:
            $ref: '#/components/schemas/DecisionRow'
        liveTradeFeed:
          type: array
          items:
            $ref: '#/components/schemas/TradeRow'
        queueSummary:
          type: array
          items:
            $ref: '#/components/schemas/QueueRow'
        recentAlerts:
          type: array
          items:
            $ref: '#/components/schemas/AlertRow'
...
    DecisionRow:
      type: object
      required:
        - correlationId
        - strategyId
        - marketTicker
        - decisionAction
        - currentLifecycleStage
        - latestEventAt
        - sourcePathMode
      properties:
        correlationId:
          type: string
        strategyId:
          type: string
        symbol:
          type: string
        marketTicker:
          type: string
        decisionAction:
          type: string
        reasonSummary:
          type: string
        currentLifecycleStage:
          type: string
        currentOutcomeStatus:
          type: string
        latestEventAt:
          type: string
          format: date-time
        sourcePathMode:
          type: string
          enum: [publisher_only, direct_only, hybrid]
        degraded:
          type: boolean
    TradeRow:
      type: object
      required:
        - correlationId
        - tradeAttemptKey
        - strategyId
        - marketTicker
        - status
        - latestSeenAt
        - sourcePathMode
      properties:
        correlationId:
          type: string
        tradeAttemptKey:
          type: string
        strategyId:
          type: string
        symbol:
          type: string
        marketTicker:
          type: string
        status:
          type: string
        publishStatus:
          type: string
          nullable: true
        lastResultStatus:
          type: string
          nullable: true
        latestSeenAt:
          type: string
          format: date-time
        sourcePathMode:
          type: string
          enum: [publisher_only, direct_only, hybrid]
        degraded:
          type: boolean
...
    AlertRow:
      type: object
      required:
        - alertId
        - alertType
        - severity
        - status
        - summary
        - latestSeenAt
        - detailPath
      properties:
        alertId:
          type: string
        alertType:
          type: string
        severity:
          type: string
        status:
          type: string
        summary:
          type: string
        componentType:
          type: string
        componentKey:
          type: string
        latestSeenAt:
          type: string
          format: date-time
        detailPath:
          type: string
...
    PnlSummaryCard:
      type: object
      required:
        - scopeType
        - scopeKey
        - realizedPnlNet
        - unrealizedPnlNet
        - feesTotal
      properties:
        scopeType:
          type: string
        scopeKey:
          type: string
        realizedPnlNet:
          type: number
        unrealizedPnlNet:
          type: number
        feesTotal:
          type: number
        stale:
          type: boolean
        partial:
          type: boolean
        freshnessTimestamp:
          type: string
          format: date-time
          nullable: true
        disagreementCount:
          type: integer
    QueueRow:
      type: object
      required: [componentName, queueName, messageCount, consumerCount, sampledAt]
      properties:
        componentName:
          type: string
        queueName:
          type: string
        messageCount:
          type: integer
        messagesReady:
          type: integer
        messagesUnacknowledged:
          type: integer
        consumerCount:
          type: integer
        oldestMessageAgeSeconds:
          type: number
          nullable: true
        dlqMessageCount:
          type: integer
          nullable: true
        dlqGrowthTotal:
          type: integer
          nullable: true
        reconnectStatus:
          type: string
          nullable: true
        sampledAt:
          type: string
          format: date-time
...
    SystemHealthSummary:
      type: object
      required: [status, freshnessTimestamp]
      properties:
        status:
          type: string
        freshnessTimestamp:
          type: string
          format: date-time
        degraded:
          type: boolean
...
    AlertRuleListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AlertRuleConfig'
    AlertRuleConfig:
      type: object
      required:
        - alertRuleId
        - alertType
        - severity
        - thresholdValue
        - thresholdUnit
        - enabled
        - version
      properties:
        alertRuleId:
          type: string
        alertType:
          type: string
        scopeType:
          type: string
        scopeKey:
          type: string
        severity:
          type: string
        comparisonOperator:
          type: string
        thresholdValue:
          type: number
        thresholdUnit:
          type: string
        evaluationWindowSeconds:
          type: integer
        consecutiveFailuresRequired:
          type: integer
        cooldownSeconds:
          type: integer
        enabled:
          type: boolean
        configSource:
          type: string
        version:
          type: integer
    AlertRuleUpdateRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
        severity:
          type: string
        thresholdValue:
          type: number
        thresholdUnit:
          type: string
        evaluationWindowSeconds:
          type: integer
        consecutiveFailuresRequired:
          type: integer
        cooldownSeconds:
          type: integer
        enabled:
          type: boolean
    AlertRuleMutationResponse:
      type: object
      required: [rule, audit]
      properties:
        rule:
          $ref: '#/components/schemas/AlertRuleConfig'
        audit:
          $ref: '#/components/schemas/MutationAuditInfo'
...
    MutationAuditInfo:
      type: object
      required: [auditLogId, result, occurredAt]
      properties:
        auditLogId:
          type: string
        result:
          type: string
          enum: [succeeded, failed]
        occurredAt:
          type: string
          format: date-time
        actorUserId:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ConflictResponse:
      type: object
      required: [code, message, currentVersion]
      properties:
        code:
          type: string
        message:
          type: string
        currentVersion:
          type: integer
    ValidationErrorResponse:
      type: object
      required: [code, message, issues]
      properties:
        code:
          type: string
        message:
          type: string
        issues:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
    ValidationIssue:
      type: object
      required: [path, message]
      properties:
        path:
          type: string
        message:
          type: string